requirements.txt
asgiref==3.4.1
Django==3.2.7
django-countries==7.2.1
django-crispy-forms==1.12.0
django-extensions==3.1.3
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-datatables==0.7.0
Pillow==8.3.1
pkg-resources==0.0.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.2
